// src/services/generated/model/invoiceItem.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * Lexxadata Customer Scraper API
 * A structured API to search and scrape customer data from the NMS portal.
 * OpenAPI spec version: 1.0.0
 */
import type { InvoiceItemStatus } from './invoiceItemStatus';
import type { InvoiceItemPackage } from './invoiceItemPackage';
import type { InvoiceItemPeriod } from './invoiceItemPeriod';
import type { InvoiceItemMonth } from './invoiceItemMonth';
import type { InvoiceItemYear } from './invoiceItemYear';
import type { InvoiceItemPaymentLink } from './invoiceItemPaymentLink';
import type { InvoiceItemDescription } from './invoiceItemDescription';

export interface InvoiceItem {
  status?: InvoiceItemStatus;
  package?: InvoiceItemPackage;
  period?: InvoiceItemPeriod;
  month?: InvoiceItemMonth;
  year?: InvoiceItemYear;
  payment_link?: InvoiceItemPaymentLink;
  description?: InvoiceItemDescription;
}